Add Full Reference Guide and extra links to the Quick Reference Guide
This commit only extends the FitNesse User Guide.
There is  a wealth of information but it is difficult to find as no full
table of content exist. Such a TOC named "FullRefernceGuide" has been
added and can be accessed from the User Guide page and the footer on
every page of the user guide.
The Quick Reference Guide had some brooken links due to the recent
refactoring. These have ben fixed. Some more links have been added to
give more details to the user if the information on the quck reference
guide is hard to understand for a new FitNesse user.