StackTraceEnricherTest fails under Windows if folder contains spaces
This fixes a problem with this junit test under windows if a folder name
contains a space. In the identified location the space is replaced with
"%20" will classpath has spaces.  This fixes this problem by replacing
all "%20" with a space in the location. As the test already replaces /
and \ one more replacement should not break anything further.