Smart Search Screen
Search Results screen show now the search input form as well.
If the user didn't found what he was looking for he can modify the
search and start again.
Search screen now also allows to define search scope: the area of the
wiki that will be included in the search