859: Change behavior for loading classes by graceful name
When the class is not found, try to load the class by changing the case
of the first letter of the classname. This helps fallback scenarios
where the Disgracer class does not give the correct name.

I.E.:
|Table:rest|...|
|...|

This will look for a class 'rest', while we want to find the class
'Rest'.