Fix localization failures regarding dates in tests
I faced a localization issue while running the tests and it has to do
with the format of the dates.  In my system, for example, one test would
produce an HTML page that would contain:

   <td>Fri Dec 05 01:20:00 GMT+01:00 1980</td>

If I run the tests, though, with setting TZ=UTC for example, the
formatted page won't contain the timezone offset:

   <td>Fri Dec 05 01:20:00 UTC 1980</td>

We could fix this problem by either using appropriate TZ settings at the
tests or by hard-coding the date format to be UTC-based, for example.

Since it appears that the exact format is not that important, we can
simply change the tests to also count for a timezone offset in the
resulted output.